Refactor printedpad
