+++ keyboards/printedpad/post_rules.mk
@@ -1,0 +1,3 @@
+ifeq ($(strip $(OLED_ENABLE)), yes)
+    SRC += oled.c
+endif
