--- keyboards/printedpad/rules.mk
@@ -1,1 +1,0 @@
-SRC += oled.c
